BSN
In the present-day scenario, we find a large number of elderly people staying alone in flats or at isolated places. The present system of a patient physically moving to a healthcare provider has yielded no efficient result. Therefore, an attempt was made to make use of modern developed technologies like IOT and Body Sensor Network (BSN). In this project, therefore it was thought to make use of the already developed technologies like IOT and BSN. However, development of this technology demands careful consideration of security of patient. An attempt is therefore made to cover major security requirements. Subsequently a secure IOT based healthcare system using various body sensors already available in the market is proposed. The proposed system envisages use of various body sensors which can be worn by the patient and can be easily managed by the veteran patients. The output of the sensors used are consolidated by a local processing unit which further forward it to the healthcare server through the mobile network. The project uses NodeMCU (ESP8266) worn by the patient on his wrist and a few body sensors which are commercially available in the market. The front end of the system consists of Arduino IDE and Blynk and Thingspeak an online analytical IoT tool act as back end of the system. The proposed system enables a patient’s medical status to be monitored by family member or friend on real time and thus a patient getting timely and quality healthcare. This system when further developed will be definitely a blessing particularly to isolated elders suffering from prolonged life style diseases. About 75 % of elders of today is suffering from one or two of the life style diseases namely diabetes, hypertension etc. The proposed BSN healthcare system come to the timely rescue of these elderly patients who are otherwise not able to make it to hospital due to poor health condition. Depending on a patient’s requirement, he is made to wear various body sensors. The proposed body sensor network with the help of IOT make it possible for the designated healthcare team to get the patient health condition instantly. Components used in the proposed system are commercially available and are affordable. BSN healthcare system with the help of IOT is a very valuable instrument for healthcare and is definitely going to revolutionize the healthcare system of the world.

 